restructured app so it's not all one file
