--- app.py
@@ -1,505 +1,0 @@
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
-from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user, current_user
-from flask_socketio import SocketIO, emit, join_room, leave_room
-from werkzeug.security import generate_password_hash, check_password_hash
-from flask_sqlalchemy import SQLAlchemy
-from flask_mail import Mail, Message
-from itsdangerous import URLSafeTimedSerializer, SignatureExpired
-from config import Config
-import secrets
-import datetime
-from collections import defaultdict
-
-app = Flask(__name__)
-app.config.from_object(Config)
-
-db = SQLAlchemy(app)
-socketio = SocketIO(app)
-mail = Mail(app)
-login_manager = LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = 'login'
-
-# Token serializer for reset links
-serializer = URLSafeTimedSerializer(app.config['SECRET_KEY'])
-
-# Update the active users data structure to be indexed by room
-active_rooms = defaultdict(lambda: {
-    'users': {},  # Dictionary of username -> user data for each room
-    'created_at': datetime.datetime.now()
-})
-
-# User data will look like:
-# active_rooms[room_id]['users'][username] = {
-#     'sid': session_id,
-#     'last_heartbeat': timestamp,
-#     'joined_at': timestamp
-# }
-
-
-# API Key model
-class ApiKey(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    key = db.Column(db.String(64), unique=True, nullable=False)
-    name = db.Column(db.String(50), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.datetime.now(datetime.UTC))
-    last_used = db.Column(db.DateTime)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-
-    def __init__(self, user_id, name):
-        self.key = secrets.token_urlsafe(32)
-        self.user_id = user_id
-        self.name = name
-
-
-class ChatRoom(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(50), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.datetime.now(datetime.UTC))
-    creator_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    messages = db.relationship('ChatMessage', backref='room', lazy=True, cascade='all, delete-orphan')
-
-
-class ChatMessage(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    content = db.Column(db.String(500), nullable=False)
-    timestamp = db.Column(db.DateTime, default=datetime.datetime.now(datetime.UTC))
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    room_id = db.Column(db.Integer, db.ForeignKey('chat_room.id'), nullable=False)
-    sender = db.relationship('User', backref=db.backref('messages_sent', lazy=True))
-
-    def to_dict(self):
-        return {
-            'content': self.content,
-            'timestamp': self.timestamp.strftime('%Y-%m-%d %H:%M:%S'),
-            'username': self.sender.username
-        }
-
-
-class User(UserMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password_hash = db.Column(db.String(120), nullable=False)
-    rooms_created = db.relationship('ChatRoom', backref='creator', lazy=True)
-    api_keys = db.relationship('ApiKey', backref='user', lazy=True)
-
-    def set_password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def check_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-
-@login_manager.user_loader
-def load_user(user_id):
-    return db.session.get(User, int(user_id))
-
-
-def auth_with_api_key():
-    api_key = request.headers.get('X-API-Key')
-    if not api_key:
-        return None
-
-    key = ApiKey.query.filter_by(key=api_key).first()
-    if key:
-        key.last_used = datetime.datetime.now(datetime.UTC)
-        db.session.commit()
-        return key.user
-    return None
-
-
-# Room management routes
-@app.route('/rooms')
-@login_required
-def rooms():
-    available_rooms = ChatRoom.query.all()
-    return render_template('rooms.html', rooms=available_rooms)
-
-
-@app.route('/room/<int:room_id>')
-@login_required
-def chat_room(room_id):
-    room = ChatRoom.query.get_or_404(room_id)
-    messages = ChatMessage.query.filter_by(room_id=room_id).order_by(ChatMessage.timestamp.desc()).limit(50).all()
-    return render_template('chat.html', room=room, messages=messages)
-
-
-@app.route('/room/create', methods=['POST'])
-@login_required
-def create_room():
-    name = request.form.get('room_name')
-    if not name:
-        flash('Room name is required')
-        return redirect(url_for('rooms'))
-
-    room = ChatRoom(name=name, creator_id=current_user.id)
-    db.session.add(room)
-    db.session.commit()
-
-    flash(f'Room "{name}" created successfully')
-    return redirect(url_for('chat_room', room_id=room.id))
-
-
-@app.route('/room/<int:room_id>/delete', methods=['POST'])
-@login_required
-def delete_room(room_id):
-    room = ChatRoom.query.get_or_404(room_id)
-    if room.creator_id != current_user.id:
-        flash('You can only delete rooms you created')
-        return redirect(url_for('rooms'))
-
-    db.session.delete(room)
-    db.session.commit()
-    flash('Room deleted successfully')
-    return redirect(url_for('rooms'))
-
-
-# WebSocket events
-@socketio.on('connect')
-def handle_connect():
-    if current_user.is_authenticated:
-        emit('connection_response', {'status': 'connected'})
-
-
-@socketio.on('join_room')
-def handle_join_room(data):
-    if current_user.is_authenticated:
-        room_id = data['room_id']
-
-        # Remove user from any previous room
-        for r_id in active_rooms:
-            if current_user.username in active_rooms[r_id]['users']:
-                leave_room(f"room_{r_id}")
-                active_rooms[r_id]['users'].pop(current_user.username)
-                # Notify previous room that user left
-                emit('user_left', {
-                    'username': current_user.username,
-                    'room_id': r_id
-                }, to=f"room_{r_id}")
-                # Update previous room's user list
-                emit('room_users_update', {
-                    'users': list(active_rooms[r_id]['users'])
-                }, to=f"room_{r_id}")
-                print('join_room cleanup')
-                print(list(active_rooms[room_id]['users']))
-
-        # Join new room
-        join_room(f"room_{room_id}")
-        active_rooms[room_id]['users'][current_user.username] = {
-            'sid': request.sid,
-            'last_heartbeat': datetime.datetime.now(),
-            'joined_at': datetime.datetime.now()
-        }
-
-        # Notify room of new user
-        emit('user_joined', {
-            'username': current_user.username,
-            'room_id': room_id
-        }, to=f"room_{room_id}")
-
-        # Send updated user list for this room
-        emit('room_users_update', {
-            'users': list(active_rooms[room_id]['users'])
-        }, to=f"room_{room_id}")
-        print('join_room')
-        print(list(active_rooms[room_id]['users']))
-
-
-@socketio.on('leave_room')
-def handle_leave_room(data):
-    if current_user.is_authenticated:
-        room_id = data['room_id']
-        if room_id in active_rooms and current_user.username in active_rooms[room_id]['users']:
-            leave_room(f"room_{room_id}")
-            active_rooms[room_id]['users'].pop(current_user.username)
-
-            # Notify room that user left
-            emit('user_left', {
-                'username': current_user.username,
-                'room_id': room_id
-            }, to=f"room_{room_id}")
-
-            # Update room's user list
-            emit('room_users_update', {
-                'users': list(active_rooms[room_id]['users'])
-            }, to=f"room_{room_id}")
-
-            # Clean up empty rooms
-            if not active_rooms[room_id]['users']:
-                active_rooms.pop(room_id)
-            print('leave_room')
-            print(list(active_rooms[room_id]['users']))
-
-
-@socketio.on('heartbeat')
-def handle_heartbeat(data):
-    if current_user.is_authenticated:
-        for room_id in active_rooms:
-            if current_user.username in active_rooms[room_id]['users']:
-                active_rooms[room_id]['users'][current_user.username]['last_heartbeat'] = datetime.datetime.now()
-                print('doki')
-                print(data)
-                check_inactive_users(data['room_id'])
-
-
-@socketio.on('disconnect')
-def handle_disconnect():
-    if current_user.is_authenticated:
-        for room_id in list(active_rooms):  # Use list to avoid runtime modification issues
-            room_data = active_rooms[room_id]
-            if current_user.username in room_data['users']:
-                room_data['users'].pop(current_user.username)
-
-                # Notify room that user left
-                emit('user_left', {
-                    'username': current_user.username,
-                    'room_id': room_id
-                }, to=f"room_{room_id}")
-
-                # Update room's user list
-                emit('room_users_update', {
-                    'users': list(active_rooms[room_id]['users'])
-                }, to=f"room_{room_id}")
-
-                # Clean up empty rooms
-                if not active_rooms[room_id]['users']:
-                    active_rooms.pop(room_id)
-                print('disconnect')
-                print(list(active_rooms[room_id]['users']))
-
-
-@socketio.on('send_message')
-def handle_message(data):
-    if current_user.is_authenticated:
-        # Find which room the user is in
-        room_id = None
-        for r_id in active_rooms:
-            if current_user.username in active_rooms[r_id]['users']:
-                room_id = r_id
-                break
-
-        if room_id:
-            message = ChatMessage(
-                content=data['message'],
-                user_id=current_user.id,
-                room_id=room_id
-            )
-            db.session.add(message)
-            db.session.commit()
-
-            emit('new_message', {
-                'username': current_user.username,
-                'message': message.content,
-                'timestamp': message.timestamp.strftime('%Y-%m-%d %H:%M:%S')
-            }, to=f"room_{room_id}")
-
-
-# Background task to check for inactive users
-def check_inactive_users(room_id):
-    current_time = datetime.datetime.now()
-    inactive_threshold = datetime.timedelta(minutes=1)
-
-    for username in list(active_rooms[room_id]['users']):
-        last_heartbeat = active_rooms[room_id]['users'][username]['last_heartbeat']
-        if current_time - last_heartbeat > inactive_threshold:
-            # User has been inactive, remove them from the room
-            active_rooms[room_id]['users'].pop(username)
-
-            emit('user_left', {
-                'username': username,
-                'room_id': room_id
-            }, to=f"room_{room_id}")
-
-            emit('room_users_update', {
-                'users': list(active_rooms[room_id]['users'])
-            }, to=f"room_{room_id}")
-
-            # Clean up empty rooms
-            if not active_rooms[room_id]['users']:
-                active_rooms.pop(room_id)
-            print('check_inactive_users')
-            print(list(active_rooms[room_id]['users']))
-
-
-# API Routes
-@app.route('/api/test', methods=['GET'])
-def api_test():
-    user = auth_with_api_key()
-    if not user:
-        return jsonify({'error': 'Invalid API key'}), 401
-    return jsonify({'message': f'Hello {user.username}! Your API key is valid.'})
-
-
-# Web Routes
-@app.route('/')
-def index():
-    return render_template('index.html')
-
-
-@app.route('/chat')
-@login_required
-def chat():
-    messages = ChatMessage.query.order_by(ChatMessage.timestamp.desc()).limit(50).all()
-    return render_template('chat.html', messages=messages)
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('dashboard'))
-
-    if request.method == 'POST':
-        username = request.form['username']
-        password = request.form['password']
-        remember = 'remember' in request.form  # Check if remember-me checkbox is checked
-
-        user = User.query.filter_by(username=username).first()
-
-        if user and user.check_password(password):
-            login_user(user, remember=remember)
-            return redirect(url_for('dashboard'))
-        else:
-            flash('Invalid username or password')
-
-    return render_template('login.html')
-
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('dashboard'))
-
-    if request.method == 'POST':
-        username = request.form['username']
-        email = request.form['email']
-        password = request.form['password']
-
-        # Check if username or email already exists
-        if User.query.filter_by(username=username).first():
-            flash('Username already exists')
-            return redirect(url_for('register'))
-
-        if User.query.filter_by(email=email).first():
-            flash('Email already registered')
-            return redirect(url_for('register'))
-
-        # Create new user
-        user = User(username=username, email=email)
-        user.set_password(password)
-        db.session.add(user)
-        db.session.commit()
-
-        flash('Registration successful')
-        return redirect(url_for('login'))
-
-    return render_template('register.html')
-
-
-@app.route('/dashboard')
-@login_required
-def dashboard():
-    return render_template('dashboard.html')
-
-
-@app.route('/api-keys')
-@login_required
-def api_keys():
-    return render_template('api_keys.html', api_keys=current_user.api_keys)
-
-
-@app.route('/api-keys/generate', methods=['POST'])
-@login_required
-def generate_api_key():
-    key_name = request.form.get('key_name', 'New API Key')
-    api_key = ApiKey(user_id=current_user.id, name=key_name)
-    db.session.add(api_key)
-    db.session.commit()
-    flash(f'New API key generated: {api_key.key}')
-    return redirect(url_for('api_keys'))
-
-
-@app.route('/api-keys/<int:key_id>/delete', methods=['POST'])
-@login_required
-def delete_api_key(key_id):
-    api_key = ApiKey.query.get_or_404(key_id)
-    if api_key.user_id != current_user.id:
-        flash('Unauthorized')
-        return redirect(url_for('api_keys'))
-
-    db.session.delete(api_key)
-    db.session.commit()
-    flash('API key deleted')
-    return redirect(url_for('api_keys'))
-
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-
-def send_reset_email(user):
-    # Generate timed token
-    token = serializer.dumps(user.email, salt='password-reset-salt')
-
-    # Create reset link
-    reset_url = url_for('reset_password', token=token, _external=True)
-
-    # Send email
-    msg = Message('Password Reset Request',
-                      sender=app.config['MAIL_USERNAME'],
-                      recipients=[user.email])
-
-    msg.body = f'''To reset your password, visit the following link:
-{reset_url}
-
-If you did not make this request, please ignore this email.
-'''
-    mail.send(msg)
-
-
-@app.route('/forgot_password', methods=['GET', 'POST'])
-def forgot_password():
-    if request.method == 'POST':
-        email = request.form['email']
-        user = User.query.filter_by(email=email).first()
-
-        if user:
-            send_reset_email(user)
-            flash('An email has been sent with instructions to reset your password.')
-            return redirect(url_for('login'))
-        else:
-            flash('Email address not found')
-
-    return render_template('forgot_password.html')
-
-
-@app.route('/reset_password/<token>', methods=['GET', 'POST'])
-def reset_password(token):
-    try:
-        # Verify token and get email (expires after 1 hour)
-        email = serializer.loads(token, salt='password-reset-salt', max_age=3600)
-    except SignatureExpired:
-        flash('The password reset link has expired.')
-        return redirect(url_for('forgot_password'))
-
-    if request.method == 'POST':
-        user = User.query.filter_by(email=email).first()
-
-        if user:
-            new_password = request.form['password']
-            user.set_password(new_password)
-            db.session.commit()
-            flash('Your password has been updated! Please login.')
-            return redirect(url_for('login'))
-
-    return render_template('reset_password.html')
-
-
-if __name__ == '__main__':
-    with app.app_context():
-        #db.drop_all()     # This drops all tables
-        db.create_all()
-    app.run(debug=True)
