+++ authflask/models/__init__.py
@@ -1,0 +1,6 @@
+from .user import User
+from .chat_message import ChatMessage
+from .chat_room import ChatRoom
+from .api_key import ApiKey
+
+__all__ = ['User', 'ChatMessage', 'ChatRoom', 'ApiKey']
