+++ authflask/models/chat_message.py
@@ -1,0 +1,17 @@
+from datetime import datetime, UTC
+from authflask.extensions import db
+
+class ChatMessage(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    content = db.Column(db.String(500), nullable=False)
+    timestamp = db.Column(db.DateTime, default=datetime.now(UTC))
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    room_id = db.Column(db.Integer, db.ForeignKey('chat_room.id'), nullable=False)
+    sender = db.relationship('User', backref=db.backref('messages_sent', lazy=True))
+
+    def to_dict(self):
+        return {
+            'content': self.content,
+            'timestamp': self.timestamp.strftime('%Y-%m-%d %H:%M:%S'),
+            'username': self.sender.username
+        }
