+++ authflask/models/user.py
@@ -1,0 +1,17 @@
+from flask_login import UserMixin
+from werkzeug.security import generate_password_hash, check_password_hash
+from authflask.extensions import db
+
+class User(UserMixin, db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    password_hash = db.Column(db.String(120), nullable=False)
+    rooms_created = db.relationship('ChatRoom', backref='creator', lazy=True)
+    api_keys = db.relationship('ApiKey', backref='user', lazy=True)
+
+    def set_password(self, password):
+        self.password_hash = generate_password_hash(password)
+
+    def check_password(self, password):
+        return check_password_hash(self.password_hash, password)
