+++ authflask/templates/api_keys.html
@@ -4,7 +4,7 @@
     <h2>API Keys</h2>
     
     <div class="mb-4">
-        <form method="POST" action="{{ url_for('generate_api_key') }}" class="row g-3">
+        <form method="POST" action="{{ url_for('api_keys.generate_api_key') }}" class="row g-3">
             <div class="col-auto">
                 <input type="text" class="form-control" name="key_name" placeholder="API Key Name" required>
             </div>
@@ -33,7 +33,7 @@
                     <td>{{ key.created_at.strftime('%Y-%m-%d %H:%M UTC') }}</td>
                     <td>{{ key.last_used.strftime('%Y-%m-%d %H:%M UTC') if key.last_used else 'Never' }}</td>
                     <td>
-                        <form method="POST" action="{{ url_for('delete_api_key', key_id=key.id) }}" style="display: inline;">
+                        <form method="POST" action="{{ url_for('api_keys.delete_api_key', key_id=key.id) }}" style="display: inline;">
                             <button type="submit" class="btn btn-danger btn-sm" onclick="return confirm('Are you sure you want to delete this API key?')">Delete</button>
                         </form>
                     </td>
