+++ authflask/templates/base.html
@@ -9,7 +9,7 @@
 <body>
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
         <div class="container">
-            <a class="navbar-brand" href="{{ url_for('index') }}">Chat App</a>
+            <a class="navbar-brand" href="{{ url_for('main.index') }}">Chat App</a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
                 <span class="navbar-toggler-icon"></span>
             </button>
@@ -17,10 +17,10 @@
                 <ul class="navbar-nav">
                     {% if current_user.is_authenticated %}
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('rooms') }}">Chat Rooms</a>
+                            <a class="nav-link" href="{{ url_for('main.rooms') }}">Chat Rooms</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('api_keys') }}">API Keys</a>
+                            <a class="nav-link" href="{{ url_for('main.api_keys') }}">API Keys</a>
                         </li>
                     {% endif %}
                 </ul>
@@ -30,14 +30,14 @@
                             <span class="nav-link text-muted">{{ current_user.username }}</span>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
+                            <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout</a>
                         </li>
                     {% else %}
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('login') }}">Login</a>
+                            <a class="nav-link" href="{{ url_for('auth.login') }}">Login</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="{{ url_for('register') }}">Register</a>
+                            <a class="nav-link" href="{{ url_for('auth.register') }}">Register</a>
                         </li>
                     {% endif %}
                 </ul>
