+++ authflask/templates/chat.html
@@ -12,7 +12,7 @@
                 </div>
                 <div>
                     <span id="connection-status" class="badge bg-secondary">Disconnected</span>
-                    <a href="{{ url_for('rooms') }}" class="btn btn-outline-secondary btn-sm ml-2">Exit Room</a>
+                    <a href="{{ url_for('main.rooms') }}" class="btn btn-outline-secondary btn-sm ml-2">Exit Room</a>
                 </div>
             </div>
             <div class="card-body">
