+++ authflask/templates/forgot_password.html
@@ -12,7 +12,7 @@
                 <button type="submit" class="btn btn-primary">Request Password Reset</button>
             </form>
             <div class="mt-3">
-                <a href="{{ url_for('login') }}">Back to Login</a>
+                <a href="{{ url_for('auth.login') }}">Back to Login</a>
             </div>
         </div>
     </div>
