+++ authflask/templates/register.html
@@ -20,7 +20,7 @@
                 <button type="submit" class="btn btn-primary">Register</button>
             </form>
             <div class="mt-3">
-                <a href="{{ url_for('login') }}">Already have an account? Login</a>
+                <a href="{{ url_for('auth.login') }}">Already have an account? Login</a>
             </div>
         </div>
     </div>
