+++ authflask/templates/reset_password.html
