+++ authflask/templates/rooms.html
@@ -9,7 +9,7 @@
                 Create New Room
             </div>
             <div class="card-body">
-                <form method="POST" action="{{ url_for('create_room') }}" class="row g-3">
+                <form method="POST" action="{{ url_for('rooms.create_room') }}" class="row g-3">
                     <div class="col-auto">
                         <input type="text" class="form-control" name="room_name" placeholder="Room Name" required>
                     </div>
@@ -41,10 +41,10 @@
                                     </small>
                                 </div>
                                 <div class="btn-group" role="group">
-                                    <a href="{{ url_for('chat_room', room_id=room.id) }}" 
+                                    <a href="{{ url_for('rooms.chat_room', room_id=room.id) }}"
                                        class="btn btn-primary btn-sm">Join</a>
                                     {% if current_user.id == room.creator_id %}
-                                        <form method="POST" action="{{ url_for('delete_room', room_id=room.id) }}" 
+                                        <form method="POST" action="{{ url_for('rooms.delete_room', room_id=room.id) }}"
                                               class="d-inline"
                                               onsubmit="return confirm('Are you sure you want to delete this room?')">
                                             <button type="submit" class="btn btn-danger btn-sm">Delete</button>
